Navigation header styling
Properly styled the navigation header, it now properly sticks when scrolling and also has its navigation buttons properly animated so that it's clear that they are clickable, the page you are on won't have this animation. Spacing has also been introduced and font weight now indicated on which page you are.

- Navigation header now sticks as you scroll down the page.
- Navigation buttons now have a weight to indicate on which page you are.
- Navigation buttons now have a transition to indicate that you can click on them.
- Dark background has been removed.