Navigation header transitions
Made the navigation header change depending on page position. If it's not at the top of the page, it will get a see-through background of the contrast-color-dark custom color. If it's hovered upon after that, the see through effect will turn to the contrast-color-dark custom color altogether. Also the header navigation buttons are now slightly transparent unless they are either active or hovered upon.

-  Added two new variables for the hidden and see-through color effects of the contrast-color-dark custom color.
- Added a Javascript file that handles scroll data transformations and ties this to the html document for css access.
- When the scroll position of the page is at the top of the page, the contrast-color-dark-hidden background color will be used for the navigation bar.
- When the scroll position of the page is not at the top of the page, the contrast-color-dark-seethrough background color will be used for the navigation bar.
- When the scroll position of the page is not at the top of the page and the navigation bar is hovered over, then the contrast-color-dark background color will be used.
- Changed inactive header navigation buttons to have an opacity of .5 unless hovered upon.