Anime.js solution
Created an anime.js library solution for the 'popping' height of the articles when they're being opened. This wouldn't work with traditional html & css. The animation now consists out of two phases, phase 1 is increasing the width of the article and the second phase increases the article's height.

This solution required me to add the anime.js library to my project. I had to rework transitions within the existing animations because they were hampering functionality of the current solution. I had to remove some size settings from the CSS as well and other small styling things. The bottom padding of the articles has been replaced with an equally sized bottom border with the same color as the article background in order to create an overlap for otherwise overflowing content from the content section.

- Implemented the anime.js library into the portfolio.
- Created an article animation using anime.js that wasn't previously possible with traditional HTML/CSS.
- Removed/Changed previous css transitions that would hamper the replacement transitions.